requirements.txt
altgraph==0.17
ansicon==1.89.0
blessed==1.17.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
future==0.18.2
halo==0.0.31
idna==2.10
inquirer==2.7.0
jinxed==1.0.1
log-symbols==0.0.14
pefile==2019.4.18
pychalk==2.0.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
python-dateutil==2.8.1
python-editor==1.0.4
pywin32-ctypes==0.2.0
readchar==2.0.1
requests==2.25.1
six==1.15.0
spinners==0.0.24
termcolor==1.1.0
urllib3==1.26.3
wcwidth==0.2.5
